+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - ProxQP: bump minimum supported version to 0.2.9
+- Solver lists (available, dense, sparse) are now sets
 
 ## [2.6.0] - 2022/11/14